% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{abind_draws}
\alias{abind_draws}
\title{Bind a list of \verb{[n x t x t]} draw arrays along the draw dimension}
\usage{
abind_draws(...)
}
\description{
Bind a list of \verb{[n x t x t]} draw arrays along the draw dimension
}
\keyword{internal}
