% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{.eta_index}
\alias{.eta_index}
\title{ETA index (position in the BGLR ETA list) of a term key}
\usage{
.eta_index(fit, key)
}
\description{
ETA index (position in the BGLR ETA list) of a term key
}
\keyword{internal}
