% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{.flatten_interaction}
\alias{.flatten_interaction}
\title{Flatten an interaction expression (\code{a:b:c}) into its components}
\usage{
.flatten_interaction(e)
}
\description{
Flatten an interaction expression (\code{a:b:c}) into its components
}
\keyword{internal}
