% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.R
\name{.fw_find_key}
\alias{.fw_find_key}
\title{Find the first ETA key of a \code{breedRB_fit} whose metadata satisfies \code{pred}}
\usage{
.fw_find_key(fit, pred)
}
\description{
Find the first ETA key of a \code{breedRB_fit} whose metadata satisfies \code{pred}
}
\keyword{internal}
