% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.R
\name{.fw_has}
\alias{.fw_has}
\title{Does a term's metadata contain a component of the given kind?}
\usage{
.fw_has(m, kind)
}
\description{
Does a term's metadata contain a component of the given kind?
}
\keyword{internal}
