% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.R
\name{.fw_has_factor}
\alias{.fw_has_factor}
\title{Does a term contain a plain factor component on variable \code{var}?}
\usage{
.fw_has_factor(m, var)
}
\description{
Does a term contain a plain factor component on variable \code{var}?
}
\keyword{internal}
