% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.R
\name{.fw_pick}
\alias{.fw_pick}
\title{Pick the first component of a term with the given kind}
\usage{
.fw_pick(components, kind)
}
\description{
Pick the first component of a term with the given kind
}
\keyword{internal}
