% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{.heritability_mt}
\alias{.heritability_mt}
\title{Multi-trait per-trait heritability (pooled over chains)}
\usage{
.heritability_mt(fit, prob)
}
\description{
Multi-trait per-trait heritability (pooled over chains)
}
\keyword{internal}
