% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{.is_fa_term}
\alias{.is_fa_term}
\title{Is the term identified by \code{key} a factor-analytic term?}
\usage{
.is_fa_term(fit, key)
}
\description{
Is the term identified by \code{key} a factor-analytic term?
}
\keyword{internal}
