% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{.mt_res_cov}
\alias{.mt_res_cov}
\title{Per-draw residual covariance matrices (multi-trait)}
\usage{
.mt_res_cov(fit, chain = 1L)
}
\description{
Per-draw residual covariance matrices (multi-trait)
}
\keyword{internal}
