% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_rm.R
\name{.outlier_group}
\alias{.outlier_group}
\title{Grouping factor from a list of columns (or a single "all" stratum)}
\usage{
.outlier_group(group_cols, n)
}
\description{
Grouping factor from a list of columns (or a single "all" stratum)
}
\keyword{internal}
