% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{.parse_response}
\alias{.parse_response}
\title{Parse the response (LHS) of the fixed formula}
\usage{
.parse_response(fixed)
}
\description{
Supports a single trait (\code{yield ~ ...}) or several traits stacked with
\code{cbind(t1, t2, ...) ~ ...} (triggers a multi-trait fit).
}
\keyword{internal}
