% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{.pooled_yhat}
\alias{.pooled_yhat}
\title{Pooled posterior-mean fitted values, aligned to the modelled data}
\usage{
.pooled_yhat(fit)
}
\description{
Pooled posterior-mean fitted values, aligned to the modelled data
}
\keyword{internal}
