% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcomp.R
\name{.post_summary}
\alias{.post_summary}
\title{Posterior summary of a numeric vector of draws}
\usage{
.post_summary(z, prob = 0.95)
}
\description{
Posterior summary of a numeric vector of draws
}
\keyword{internal}
