% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{.random_keys}
\alias{.random_keys}
\title{Names of the random (BRR) terms in a fit, matching their on-disk file keys}
\usage{
.random_keys(fit)
}
\description{
Names of the random (BRR) terms in a fit, matching their on-disk file keys
}
\keyword{internal}
