% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_rm.R
\name{.resolve_by}
\alias{.resolve_by}
\title{Add \verb{by=} columns/vector to a list of grouping columns}
\usage{
.resolve_by(by, group_cols, data, n)
}
\description{
Add \verb{by=} columns/vector to a list of grouping columns
}
\keyword{internal}
