% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_gradient.R
\name{.rr_gradient_plots}
\alias{.rr_gradient_plots}
\title{Build the four ggplot figures for an rr_gradient() result}
\usage{
.rr_gradient_plots(x, sel_ids)
}
\description{
Build the four ggplot figures for an rr_gradient() result
}
\keyword{internal}
