% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{.varcomp_mt}
\alias{.varcomp_mt}
\title{Multi-trait variance components (pooled over chains)}
\usage{
.varcomp_mt(fit, prob)
}
\description{
Multi-trait variance components (pooled over chains)
}
\keyword{internal}
