% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{.vech_index}
\alias{.vech_index}
\title{Column (i, j) index pairs of BGLR's covariance trace layout}
\usage{
.vech_index(t)
}
\description{
BGLR writes the upper triangle of the (symmetric) covariance matrix in
row-major order: (1,1), (1,2), ..., (1,t), (2,2), ..., (t,t).
}
\keyword{internal}
