% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing helper: \code{a \%||\% b} returns \code{a} unless it is \code{NULL}}
\usage{
a \%||\% b
}
\description{
Null-coalescing helper: \code{a \%||\% b} returns \code{a} unless it is \code{NULL}
}
\keyword{internal}
