% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_posterior}
\alias{plot_posterior}
\alias{plot_posterior.breedRB_fit}
\alias{plot_posterior.breedRB_h2}
\title{Posterior density plot of variance components or heritability}
\usage{
plot_posterior(x, ...)

\method{plot_posterior}{breedRB_fit}(x, ...)

\method{plot_posterior}{breedRB_h2}(x, ...)
}
\arguments{
\item{x}{A \code{breedRB_fit} (variance components) or a \code{breedRB_h2} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Posterior density plot of variance components or heritability
}
