% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prob_within}
\alias{plot_prob_within}
\title{Heatmap of within-environment top-tail probabilities}
\usage{
plot_prob_within(post_prob, top_n = 30)
}
\arguments{
\item{post_prob}{Output from \code{prob_sup_bglr()} containing:
\itemize{
\item \code{across$bayesian_stats} with \code{ID} and \code{prob_top} for ranking,
\item \code{within$perfo} with columns \code{env}, \code{ID}, and \code{prob} for environment-specific probabilities.
}}

\item{top_n}{Integer. Number of genotypes to display (ranked by \code{prob_top}). Default = 30.}
}
\value{
A ggplot object.
}
\description{
Draws a heatmap of environment-specific probabilities for Top-N genotypes,
where the probabilities represent within-environment tail probabilities computed in \code{prob_sup_bglr()}.
}
\details{
Legendre-compatible:
\itemize{
\item Probabilities are based on environment-specific posterior predictions from the Legendre
random regression (\code{within$perfo}).
}
}
\examples{
\donttest{
p5 <- plot_prob_within(post_prob = probs, top_n = 30)
print(p5)
}

}
\seealso{
prob_sup_bglr
}
