% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability.R
\name{.vm_keys}
\alias{.vm_keys}
\title{Keys of the random terms that carry a genomic (\code{vm}) component}
\usage{
.vm_keys(fit)
}
\description{
Keys of the random terms that carry a genomic (\code{vm}) component
}
\keyword{internal}
